LayoutTests/imported/w3c/web-platform-tests/svg/animations/cyclic-syncbase-expected.txt


FAIL Test cyclic for svg animations for syncbases assert_approx_equals: expected 100 +/- 1 but got 0
